feat : 커밋 좋아요 싫어요 api 완성

+++ src/main/java/com/group/commitapp/dto/gpt/GPTRequest.java
@@ -14,19 +14,19 @@ public class GPTRequest {
 
     private String model;
     private List<Message> messages;
-    private int temperature;
+    private double temperature;
     private int maxTokens;
-    private int topP;
-    private int frequencyPenalty;
-    private int presencePenalty;
+    private double topP;
+    private double frequencyPenalty;
+    private double presencePenalty;
 
     public GPTRequest(String model
             , String prompt
-            , int temperature
+            , double temperature
             , int maxTokens
-            , int topP
-            , int frequencyPenalty
-            , int presencePenalty) {
+            , double topP
+            , double frequencyPenalty
+            , double presencePenalty) {
         this.model = model;
         this.messages = new ArrayList<>();
         this.messages.add(new Message("user",prompt));
